+++ src/xrGame/hud_item_object.cpp
@@ -21,12 +21,12 @@ bool CHudItemObject::Action				(u16 cmd, u32 flags)
 	return						(CHudItem::Action(cmd,flags));
 }
 
-void CHudItemObject::SwitchState		(u32 S)
+void CHudItemObject::SwitchState		(u8 S)
 {
 	CHudItem::SwitchState		(S);
 }
 
-void CHudItemObject::OnStateSwitch		(u32 S)
+void CHudItemObject::OnStateSwitch		(u8 S)
 {
 	CHudItem::OnStateSwitch		(S);
 }
